refactor: 채널 DTO를 request와 response 패키지로 분리
