+++ src/main/java/com/example/inflace/domain/channel/dto/request/ChannelVideoFormat.java
@@ -1,4 +1,4 @@
-package com.example.inflace.domain.channel.dto.enums;
+package com.example.inflace.domain.channel.dto.request;
 
 public enum ChannelVideoFormat {
     ALL,
