+++ src/main/java/com/example/inflace/domain/channel/dto/response/ChannelKpiResponse.java
@@ -1,4 +1,4 @@
-package com.example.inflace.domain.channel.dto;
+package com.example.inflace.domain.channel.dto.response;
 
 public record ChannelKpiResponse(
         Long totalViews,
