+++ src/main/java/com/example/inflace/domain/channel/dto/response/YoutubeDataChannelResponse.java
@@ -1,4 +1,4 @@
-package com.example.inflace.domain.channel.dto;
+package com.example.inflace.domain.channel.dto.response;
 
 import java.util.List;
